feat(communications): restore three-column MessagingPanel with real chat wiring
- Recreate src/components/communications/messaging-panel.tsx with the
  previous left-rail button menu (Chats/Contacts/Requests/Followers/Settings),
  middle list column, and right chat column design.
- Wire the panel to the existing /api/conversations and messages endpoints,
  using fetchWithCsrf for sends and MentionInput for composition.
- Replace Messenger in the Communications 'Messaging' tab; keep notifications,
  requests, and all data handlers untouched.
- Remove the now-unused messenger.tsx component.
